Move TopHeader to Header
This is a remnant of an old design where there was a BottomHeader as well
